[ new ] sized version
This is quite close to a litteral translation of Section 5.1
except that we keep a certified version of the data around
(the list of lines is stored reverse so that we get constant
time access to the last line).

However it's fairly horrible in that it's using Vec & BoundedVec
of Char. The next step is to ditch this representation by introducing:

record String= (n : Nat) : Set where
  field payload : String
        .proof  : length payload = n

record String≤ (n : Nat) : Set where
  field payload : String
        .proof  : length payload ≤ n

This should nicely erase to just bunch of strings.